+++ openfisca_web_api_preview/loader/spec.py
@@ -47,7 +47,7 @@ def get_entity_json_schema(entity, tax_benefit_system):
     else:
         properties = {}
         properties.update({
-            role.plural: {
+            role.plural or role.key: {
                 'type': 'array',
                 "items": {
                     "type": "string"